backend/requirements.txt
fastapi==0.115.13
uvicorn==0.32.0
pydantic==2.9.0
pandas==2.2.3
numpy==2.3.3
supabase==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.9
httpx==0.27.2
typer==0.9.0
